+++ tests/datetime/test_datetime.py
@@ -24,7 +24,7 @@ import pytest
 from flaky import flaky
 
 import hypothesis._settings as hs
-from hypothesis import find, given, assume, settings
+from hypothesis import find, given, assume, settings, unlimited
 from hypothesis.errors import InvalidArgument
 from tests.common.debug import minimal
 from tests.common.utils import checks_deprecated_behaviour
@@ -148,7 +148,8 @@ def test_bordering_on_a_leap_year():
     x = find(
         datetimes(min_year=2002, max_year=2005),
         lambda x: x.month == 2 and x.day == 29,
-        settings=settings(database=None, max_examples=10 ** 7, timeout=-1)
+        settings=settings(
+            database=None, max_examples=10 ** 7, timeout=unlimited)
     )
     assert x.year == 2004