+++ tests/nocover/test_collective_minimization.py
@@ -47,8 +47,7 @@ def test_can_collectively_minimize(spec):
         xs = find(
             lists(spec, min_size=n, max_size=n),
             distinct_reprs,
-            settings=settings(
-                timeout=10.0, max_examples=2000))
+            settings=settings(max_shrinks=100, max_examples=2000))
         assert len(xs) == n
         assert 2 <= len(set((map(repr, xs)))) <= 3
     except NoSuchExample: