+++ tests/pytest/test_statistics.py
@@ -33,7 +33,7 @@ def test_all_valid(x):
     pass
 
 
-@settings(timeout=0.2, min_satisfying_examples=1)
+@settings(timeout=0.2, min_satisfying_examples=1, strict=False)
 @given(integers())
 def test_slow(x):
     time.sleep(0.1)