+++ docs/changes.rst
@@ -22,6 +22,22 @@ You should generally assume that an API is internal unless you have specific
 information to the contrary.
 
 -------------------
+3.16.0 - 2017-08-04
+-------------------
+
+This release introduces a deprecation of the timeout feature. This results in
+the following changes:
+
+* Creating a settings object with an explicit timeout will emit a deprecation
+  warning.
+* If your test stops because it hits the timeout (and has not found a bug) then
+  it will emit a deprecation warning.
+* There is a new value ``unlimited`` which you can import from hypothesis.
+  ``settings(timeout=unlimited)`` will *not* cause a deprecation warning.
+* There is a new health check, ``hung_test``, which will trigger after a test
+  has been running for five minutes if it is not suppressed.
+
+-------------------
 3.15.0 - 2017-08-04
 -------------------